refactor: modernize CloneXBaseRecyclerView with ListAdapter & DiffUtil
This is my final commit of 2025. It’s been a challenging year on the job hunt, and that journey is still ongoing. I hope that 2026 brings the opportunity to join an amazing team in a remote role where I can continue my professional journey. I wish for 2026 to be a year of immense learning and personal growth.